Make random forest model pickle-able
The rfr C++ random number generator is not pickle-able but is an attribute of the random forest model. 
To pickle the model we have to delete the rng before dumping and add it back after loading. 
Sadly there is no easy and robust way to also store the internal state of the rng, so we loose it.